+++ src/yafowil/tests/__init__.py
@@ -1,6 +1,7 @@
 from __future__ import print_function
 from importlib import reload
 from node.tests import NodeTestCase
+from yafowil import button
 from yafowil import common
 from yafowil import compound
 from yafowil import datatypes
@@ -24,6 +25,7 @@ class YafowilTestCase(NodeTestCase):
         super(YafowilTestCase, self).setUp()
         factory.push_state()
         factory.clear()
+        reload(button)
         reload(common)
         reload(compound)
         reload(datatypes)
@@ -54,6 +56,7 @@ tag = Tag(lambda msg: msg)
 
 def test_suite():
     from yafowil.tests import test_base
+    from yafowil.tests import test_button
     from yafowil.tests import test_common
     from yafowil.tests import test_compound
     from yafowil.tests import test_controller
@@ -72,6 +75,7 @@ def test_suite():
     suite = unittest.TestSuite()
 
     suite.addTest(unittest.findTestCases(test_base))
+    suite.addTest(unittest.findTestCases(test_button))
     suite.addTest(unittest.findTestCases(test_common))
     suite.addTest(unittest.findTestCases(test_compound))
     suite.addTest(unittest.findTestCases(test_controller))
